LayoutTests/fast/canvas/webgl/drawElementssOutOfBounds-expected.txt

Test of drawElements with out-of-bounds parameters

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Test empty index buffer
PASS context.drawElements(context.TRIANGLES, 3, context.UNSIGNED_BYTE, 0) threw exception Error: INVALID_STATE_ERR: DOM Exception 11.
PASS context.drawElements(context.TRIANGLES, 10000, context.UNSIGNED_BYTE, 0) threw exception Error: INVALID_STATE_ERR: DOM Exception 11.
PASS context.drawElements(context.TRIANGLES, 10000000000000, context.UNSIGNED_BYTE, 0) threw exception Error: INVALID_STATE_ERR: DOM Exception 11.
PASS context.drawElements(context.TRIANGLES, 1, context.UNSIGNED_BYTE, 0) threw exception Error: INVALID_STATE_ERR: DOM Exception 11.
PASS context.drawElements(context.TRIANGLES, 0, context.UNSIGNED_BYTE, -1) threw exception Error: INVALID_STATE_ERR: DOM Exception 11.
PASS context.drawElements(context.TRIANGLES, -1, context.UNSIGNED_BYTE, 1) threw exception Error: INVALID_STATE_ERR: DOM Exception 11.
PASS context.drawElements(context.TRIANGLES, 1, context.UNSIGNED_BYTE, -1) threw exception Error: INVALID_STATE_ERR: DOM Exception 11.

Test buffer with 3 byte indexes
PASS context.drawElements(context.TRIANGLES, 3, context.UNSIGNED_BYTE, 0) is undefined
PASS context.drawElements(context.TRIANGLES, 3, context.UNSIGNED_BYTE, 2) threw exception Error: INVALID_STATE_ERR: DOM Exception 11.
PASS context.drawElements(context.TRIANGLES, 10000, context.UNSIGNED_BYTE, 0) threw exception Error: INVALID_STATE_ERR: DOM Exception 11.
PASS context.drawElements(context.TRIANGLES, 10000000000000, context.UNSIGNED_BYTE, 0) threw exception Error: INVALID_STATE_ERR: DOM Exception 11.
PASS context.drawElements(context.TRIANGLES, 0, context.UNSIGNED_BYTE, -1) threw exception Error: INVALID_STATE_ERR: DOM Exception 11.
PASS context.drawElements(context.TRIANGLES, -1, context.UNSIGNED_BYTE, 1) threw exception Error: INVALID_STATE_ERR: DOM Exception 11.
PASS context.drawElements(context.TRIANGLES, 1, context.UNSIGNED_BYTE, -1) threw exception Error: INVALID_STATE_ERR: DOM Exception 11.

PASS successfullyParsed is true

TEST COMPLETE
